Finish implementation of PauliProduct
The PauliProduct class now can handle multiplications with itself
and with PauliOperators. It can also be applied to a Register.
PauliProduct will still throw an exception if the phase from
multiplying the Paulis together is not 1. Will return a PauliSum
in the future, which will have a place to store the prefactor.