Add state-controlled option to `unitaries.U`
Multi-controlled single-qubit unitaries specified by an explicit
matrix (i.e. `unitaries.U`) can now also be controlled by a given
pattern of 0 and 1 states. Regular controls only apply the gate if
all controls are in the 1 state, the extra constructor argument
`control_pattern` now takes an iterable consisting of zeros and
ones, which specify the state the controls should be in, in order
for the unitary to be applied. This is equivalent to (but faster
than) applying a Pauli-X gate to each control that is supposed to
be in the zero state, applying a regular controlled unitary, and
applying the Pauli-X gates again.