Adapt QuEST random seed call
The new version of QuEST changed the signature of `seedQuESTDefault()`
to require the current environment. To avoid circular imports, the
`gates` module now provides a function for seeding QuEST's random
number generator, which is called upon construction of the QuEST
environment. This is still hacky and should go once pyQuEST uses
QuEST as a dynamic instead of a static library.