Add `--shallow-submodules` to recommended clone
This should be possible by setting `shallow = true` in the QuEST
section of `.gitmodules`, but for now this only works if the
submodule is pointing to the `HEAD` of the repository it includes.
Until this bug is fixed, we simply add it as a command line flag to
the recommended `clone` command, where the bug is not present.