Add faster function for setting amplitudes
In addition to the usual syntax `reg[a:b] = vals` for manually setting
amplitudes in a `Register`, which calls the QuEST API many times and
is therefore quite slow, there is a new method `set_block()`, which
directly writes to the QuEST memory and is thus, despite being single-
threaded, much faster. For now this only works on CPU and distributed
versions, not with GPUs.