Add `.matrix` property to MatrixOperator
Different from `.as_matrix()`, which returns the operator on the
Hilbert space of all qubits up to the required number, the new property
`.matrix` returns the matrix on the subspace of the target qubits. This
is identical (up to numerical precision) to the matrix passed into the
constructor via the `matrix` keyword argument.