Include error override privately
While injecting the override for error handling into the QuEST library,
the pyquest source directory was added as an include directory to QuEST
with scope `PUBLIC`. This is not really necessary and the new CMake
setup of QuEST v4 throws an error in this configuration, so the scope
is switched to `PRIVATE`.